chore(startup): move startup code from cmd/main.go into startup package
